api/embeddings/pyproject.toml
[project]
name = "embeddings"
version = "0.1.0"
description = "DINOv3 feature extraction API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litserve>=0.2.8",
    "numpy>=1.24.0",
    "pillow>=11.2.1",
    "pip>=25.3",
    "torch>=2.7.0",
    "torchvision>=0.24.1",
    "transformers>=4.51.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "requests>=2.32.5",
]
